# override mutable.HashMap#updateWith and mutable.LinkedHashMap#updateWith for performance