mutable version of https://github.com/scala/scala/pull/8032

This allows us to only hash the input key and perform the hash table lookup once, rather than twice

Mostly code cribbed from the existing methods in `mutable.HashMap`, inlined and mangled so we can avoid performing the key-hashing and table-lookup more than once. Instead we store the `foundNode` and `previousNode` (either or both of which can be `null`) and mangle the `next` pointers to add things to the table/linked-list, remove things from the table/linked-list, or just update `Node.value` in place if there was already a previous value we are updating.

For `mutable.HashMap`, removals occur in-place and additions occur in the order defined by `put0`. For `mutable.LinkedHashMap`, removals occur in place and additions occur at the tail of the linked list.

I've added a pretty thorough test suite to `HashMapTests` and `LinkedHashMapTests` exercising this code in a variety of situations:

- Various transform functions
- Hash collisions
- Adding and removing from the start, middle, and end of the linked list
- Checking that `.hashCode` is only called once per `updateWith` call

Also added benchmarks, which show a measurable improvement to the performance of both operations:

```text
BEFORE

[info] Result "scala.collection.mutable.HashMapBenchmark.updateWith":
[info]   31181.097 ±(99.9%) 26909.915 ns/op [Average]
[info]   (min, avg, max) = (29481.979, 31181.097, 32132.847), stdev = 1475.023
[info]   CI (99.9%): [4271.182, 58091.012] (assumes normal distribution)
[info] # Run complete. Total time: 00:00:34
[info] Benchmark                    (size)  (stringsOnly)  (useMissingValues)  Mode  Cnt      Score       Error  Units
[info] HashMapBenchmark.updateWith      10          false                true  avgt    3    273.904 ±    14.262  ns/op
[info] HashMapBenchmark.updateWith     100          false                true  avgt    3   2623.131 ±   109.200  ns/op
[info] HashMapBenchmark.updateWith    1000          false                true  avgt    3  31181.097 ± 26909.915  ns/op

[info] Result "scala.collection.mutable.LinkedHashMapBenchmark.updateWith":
[info]   46182.148 ±(99.9%) 61228.785 ns/op [Average]
[info]   (min, avg, max) = (43733.960, 46182.148, 50007.889), stdev = 3356.156
[info]   CI (99.9%): [≈ 0, 107410.933] (assumes normal distribution)
[info] # Run complete. Total time: 00:00:34
[info] Benchmark                          (size)  (stringsOnly)  (useMissingValues)  Mode  Cnt      Score       Error  Units
[info] LinkedHashMapBenchmark.updateWith      10          false                true  avgt    3    373.098 ±   115.872  ns/op
[info] LinkedHashMapBenchmark.updateWith     100          false                true  avgt    3   3942.097 ±   404.435  ns/op
[info] LinkedHashMapBenchmark.updateWith    1000          false                true  avgt    3  46182.148 ± 61228.785  ns/op

AFTER

[info] Result "scala.collection.mutable.HashMapBenchmark.updateWith":
[info]   29130.178 ±(99.9%) 28853.438 ns/op [Average]
[info]   (min, avg, max) = (27308.944, 29130.178, 30157.602), stdev = 1581.554
[info]   CI (99.9%): [276.740, 57983.617] (assumes normal distribution)
[info] # Run complete. Total time: 00:00:34
[info] Benchmark                    (size)  (stringsOnly)  (useMissingValues)  Mode  Cnt      Score       Error  Units
[info] HashMapBenchmark.updateWith      10          false                true  avgt    3    257.002 ±    21.512  ns/op
[info] HashMapBenchmark.updateWith     100          false                true  avgt    3   2558.933 ±   347.473  ns/op
[info] HashMapBenchmark.updateWith    1000          false                true  avgt    3  29130.178 ± 28853.438  ns/op

[info] Result "scala.collection.mutable.LinkedHashMapBenchmark.updateWith":
[info]   37820.263 ±(99.9%) 27136.733 ns/op [Average]
[info]   (min, avg, max) = (36472.498, 37820.263, 39416.194), stdev = 1487.456
[info]   CI (99.9%): [10683.531, 64956.996] (assumes normal distribution)
[info] # Run complete. Total time: 00:00:34
[info] Benchmark                          (size)  (stringsOnly)  (useMissingValues)  Mode  Cnt      Score       Error  Units
[info] LinkedHashMapBenchmark.updateWith      10          false                true  avgt    3    284.379 ±     7.107  ns/op
[info] LinkedHashMapBenchmark.updateWith     100          false                true  avgt    3   3033.440 ±   180.528  ns/op
[info] LinkedHashMapBenchmark.updateWith    1000          false                true  avgt    3  37820.263 ± 27136.733  ns/op
```

It seems `LinkedHashMap` benefits more than `HashMap`, with a 24% reduction in time vs a 6% reduction in time per operation. I expect this improvement to be greater in real world scenarios where `.hashCode` of the keys is non-trivial.

Not sure if there's a way for us to reasonably share code between the two effectively-identical unit test suites, but for now it's all just copy-pasta