@lihaoyi 2019-05-20
Sister PR for `LinkedHashMap` https://github.com/scala/scala/pull/8074