@diesalbla 2019-05-20
I think that this collides with https://github.com/scala/scala/pull/8032, which seems to be doing a similar thing. I will let @joshlemer to judge if he can retrieve his PR.