@Ichoran 2019-05-20
Aside: if I don't already have laws for updateWith in collection-laws, I should add some.  That ought to give fairly comprehensive testing.