@lihaoyi-databricks 2019-05-20
@joshlemer question: what's the point of the pre-emptive override? Isn't overriding a concrete method both source and binary compatible?