@szeiger 2019-05-21
Overriding a concrete method is not forward binary compatible when it introduces a bridge method. This should not be a problem for `updateWith` because its signature does not contain any types whose constraints are narrower than in the implementation that it overrides.