@fdietze 2019-05-22
I think the change should be benchmarked on the JVM, ScalaJS and native to see if it's really faster.