@lihaoyi 2019-09-03
Merged https://github.com/scala/scala/pull/8074 into this PR, added a test suite, benchmarks, and results. This should be ready for review @szeiger @joshlemer