@lihaoyi 2019-09-03
Not sure why MIMA is complaining about adding a private method to a private class