@lrytz 2019-09-03
`private[collection]` is public in bytecode. mima works on bytecode, it doesn't use the scala signatures / scala reflect / the compiler's symbol table. adding mima exception for qualified private definitions is common practice.