@lihaoyi 2019-09-03
@lrytz is this the correct place to put the MIMA exclude? https://github.com/scala/scala/blob/2.13.x/build.sbt#L126