@lihaoyi 2019-09-03
Ok I added the exclusion. For some reason it's also failing with `ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.convert.JavaCollectionWrappers#IteratorWrapper.asIterator")`, even though I don't think I ever touched that code. Let's see what Jenkins says