@lihaoyi-databricks 2019-09-03
Not sure what `continuous-integration/travis-ci/pr` is or why it is hanging, but `validate-main` passes?