@lihaoyi-databricks 2019-09-03
@szeiger yeah the relative different in speedup is a bit strange. Doesn't that mean that `HashMap` has *more* non-hashCode overhead than `LinkedHashMap` does, since removing one of the `hashCode` calls makes a smaller difference? I would have expected maintaining the linked list in `LinkedHashMap` to have more cost, but perhaps we're looking at the cost of maintaining the sorted per-bucket lists in `HashMap` that should pay for itself in lookup speed (that we're not measuring here)